Add documentation for Event struct
